requirements.txt
# Core ML — pinned to versions with Python 3.14 prebuilt wheels
scikit-learn==1.8.0
numpy==2.4.4
pandas==2.3.3          # river requires pandas<3; 2.3.3 has cp314 wheel
scipy==1.17.1
joblib==1.5.3

# Drift Detection — river 0.24.2 has cp314 wheel, requires pandas<3
river==0.24.2

# MLflow
mlflow>=2.13.0

# API
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0

# Monitoring
prometheus-client>=0.20.0

# Data
kaggle>=1.6.12

# Utilities
python-dotenv>=1.0.0
PyYAML>=6.0.0
loguru>=0.7.0
tqdm>=4.66.0

# Testing
pytest>=8.2.0
pytest-cov>=5.0.0
httpx>=0.27.0
